backend-python/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.1
websockets==12.0
sse-starlette==1.8.2
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.0
langchain==0.0.340
chromadb==0.4.18
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
